Create Word-Tagger.py
This program is for using a comma separated list of words containing wildcard words (marked with asterisks at the end), multi-word phrases, and single words, and searching pre-formatted CSV files to find Tweets containing words from this list and to tag them as either containing, in this case, political content or not. The particular columns you would want to include can be quite easily changed to suit your needs. I am working on cleaning and improving it to work easier for reputability.