Create datacorrecting.py
This is not a good implementation for the intended purpose, but it worked well in a time crunch. This was created to fill in information gaps concerning church names, addresses, states, etc. from a list containing all this relevant information. The program uses the church's particular Twitter handle as the key to finding all the appropriate information from the master file containing all the church's information. This way I was able to quickly populate a file with thousands of lines of missing information using one piece of relevant information as the key. The file of missing information contained only the Tweets and Twitter account handles. The rest of the information required a fill in the blank approach.